+++ src/main/java/colorful/starbucks/product/presentation/ProductController.java
@@ -3,8 +3,8 @@ package colorful.starbucks.product.presentation;
 import colorful.starbucks.common.response.ApiResponse;
 import colorful.starbucks.product.application.ProductService;
 import colorful.starbucks.product.dto.request.ProductCreateRequestDto;
-import colorful.starbucks.product.vo.ProductCreateRequestVo;
-import colorful.starbucks.product.vo.ProductCreateResponseVo;
+import colorful.starbucks.product.vo.request.ProductCreateRequestVo;
+import colorful.starbucks.product.vo.response.ProductCreateResponseVo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
