+++ src/main/java/colorful/starbucks/product/vo/response/ProductCreateResponseVo.java
@@ -1,4 +1,4 @@
-package colorful.starbucks.product.vo;
+package colorful.starbucks.product.vo.response;
 
 import lombok.Builder;
 import lombok.Getter;
